Controller struct

+++ include/mixer.h
@@ -29,19 +29,19 @@ typedef enum
 
 typedef struct
 {
-  int32_t F;
-  int32_t x;
-  int32_t y;
-  int32_t z;
+  float F;
+  float x;
+  float y;
+  float z;
 } command_t;
 
 typedef struct
 {
   output_type_t output_type[8];
-  int32_t F[8];
-  int32_t x[8];
-  int32_t y[8];
-  int32_t z[8];
+  float F[8];
+  float x[8];
+  float y[8];
+  float z[8];
 } mixer_t;
 
 extern command_t _command;
